Allow custom metrics in gobblin runtime. Refactor Kafka reporter.
